combine  MemAlloc + pthread_mutex_init => pthread_mutex_new  MemAlloc + pthread_cond_init => pthread_cond_new  pthread_mutex_destroy + MemFree => pthread_mutex_delete  pthread_cond_destroy + MemFree => pthread_cond_delete
This was upon revisiting MemAlloc and was going to change to use NEW,
however a REF ARRAY OF CHAR and its first byte are not at the same
address, which is a bit annoying, so didn't go that route.


This isn't clearly worth anything, but it also shouldn't be controversial.
  A small net reduction in the number of C symbols: since "size" and "init"
   now combine down to "new".
 MemAlloc and MemFree go away now, but the allocations are still done with calloc,
  but from the C code.
While there, treat ENOMEM like other out of memory cases.
  Any other errors including EGAIN trigger assertion failures.


Remove the avoidance of RAISE.
 Two of the uses are within a TRY, one is not.
 But the reasons are moot now, now that initialization order is better.

Rename "UNDERUNDER_THREAD" to "M3CONFIG_THREAD_LOCAL_STORAGE",
to be used shortly.