first draft of code to create Debian and IExpress packages
It is very minimal and feature-less, in particular
there is only min and std, and they do overlap, so would
require 'force' or somesuch to install. But this is easily
fixed once the code is working -- it should consume the
package breakdown Olaf recently developed.


IExpress (iexpress.exe) is a builtin hybrid console/gui-wizard
documented in MSDN, that builds self extracting installers.
It creates a .cab inside an .exe.
If you rename the .exe to .cab, Explorer does NOT understand it,
but e.g. 7-zip does.
Developer can provide an installer program that moves the files into
place and/or a post-install program and/or an initial confirmation
prompt and/or a done message, etc. Here we just have minimal for now.
Later we might probe around for the compiler/linker, or maybe even
run cminstall (but that somewhat bags for bringing back cygwin.dll/gzip.exe/tar.exe
which seems like overkill, probably better is a mode of cminstall that just
mucks the config file; but I'm actually hoping to migrate the environment
variable probing in pylib.py into NT386, that should suffice, so continue
to have no installer.)


Key developments today:
  1) Every previous attempt to learn about Debian packaging landed me in
  a quagmire of reading about source based builds. Finally I found
  the pieces of how to make something that just installs binaries that
  are already built. Furthermore, there is no need to use any dpkg* tools,
  it is just tar, gzip, ar, and optionally bzip2/lzma, very easy, can
  be created on just about any host (er, then again, not without significant
  cross build improvements..)

  2) Discovered IExpress.


Not yet tested, waiting for make-dist.py to finish to produce the input
but constituent pieces verified so I could know that I know what I'm doing..


Yet missing is RPM and BSD packaging.