An arguably better rendition of -no-m3ship-resolution, that works better than the previous version. My previous change definitely fixed many cases, due to the dotdots and ,any cases were also either still or newly broken. I didn't look closely enough at the results for the first version to know if I'd seen all the errors.
This should get it all correct.
Though it is probably riskier in that the previous rendition
avoided touching any preexisting code path.
It tediously duplicated some of the path formations and
sleazily back substituted in others.
This version always just uses sleazy back substitution, which
at least is simple and works and isn't tedious.

As well, the responsibility for unquoting is pushed down
into Unresolve, even when -no-m3ship-resolution isn't used.
That way we can easily remove the unsightly appends of empty
strings (as if anyone looks at .M3SHIP files..)

There is a way to do this without risking existing paths,
using just the simple sleazy reliable back substitution,
but it would not be able to remove the appends of empty strings.