address the problem with additional users of MxConfig
cm3 defines M3_PROFILING, SL, CR, and some other things
Other users of MxConfig do not
As a result, MxConfig kind of just gives up once
 execution tries to use those variables, returning
 NULL for the .Get() in progress


Reduction of this diff and other preexisting code of mine
may now be possible now that this problem is understand.
In particular, the setting of TARGET and
BUILD_DIR in NT386, NT386GNU, NT386MINGNU instead of
just NT386.common, and then this change to NT386, NT386GNU, NT386MINGNU


But this is ok asis.


It is tempting to put these defines in Quake but that would
violate the model perhaps.