Add newly developed script, do-cm3.CMD.  See help below.
--- HELP for do-cm3.CMD ---

do-cm3.CMD is intended to replicate on Windows 2000/XP the functionality of the
various "do-cm3-*.sh" scripts.

As such, it invokes cm3 on a group of packages.  Packages and their group
associations are defined in the PkgInfo.txt file.

The PkgInfo.txt file can be specified via the -P argument, or it is searched
for in the current directory, then the parent directory.

The package source tree is located relative to the PkgInfo.txt file (parent
folder), or in the current directory, or in the parent directory.

Usage:  do-cm3 {help} [min core std all] {-p path} {cm3args}

help    = display help, then exit.

-p path = specify location of PkgInfo.txt
          (if not specified, looks in ".\PkgInfo.txt" then "..\PkgInfo.txt")

cm3args = zero or more arguments to the cm3 builder, e.g., clean, build, ship
          Multiple arguments are possible.  They will be performed in the
          order given.
          "buildship" is shorthand for "build" followed by "ship".

===END do-cm3===
Enjoy!
--Randy Coleburn