add support for omitting unnecessary m3cc builds during regression tests:
If WORKSPACE is set in the environment and PREBUILT_CM3CG is set,
this file is just used as the build result. No further checks are
performed.

This is of course a hack, but our current resources of build servers
are heavily loaded by continually compiling the cm3cg backend, which
is not necessary in most cases. We can detect via CVS when it _is_.