newer Sun cc says: cc: Warning: -xarch=v8plus is deprecated, use -m32 -xarch=sparc instead cc: Warning: -xarch=v9 is deprecated, use -m64 to create 64-bit programs
probe at first use of cc (or ld since it is cc) for the warning
and use the new form if it is found

In general introduce hooks:
  configure_assembler()
  configure_c_compiler()
  configure_linker()

That are called before uses of SYSTEM_ASM, SYSTEM_CC, SYSTEM_LD.
configure_assembler was already in use for I386_DARWIN.

Even if folks really object to running this stuff as often as it is done,
the configure code can probably be used asis (say, at install time)
and the results recorded.

Personally I don't observe this stuff to be slow and it leaves the
system more flexible -- no need to reinstall/reconfigure Modula-3 if in
fact this other stuff can be changed after it is installed.