change:
InnerWait:
  ..
  Leave(giant)
  UnlockMutex:
    TlsGetValue()
    Enter(giant)
      do stuff
    Leave(giant)

to:

InnerWait:
  ..
  UnlockMutex:
    TlsGetValue()
    Enter(giant) (recursive now)
      do stuff
    Leave(giant)
  Leave(giant)


This increases the critical section
ever so slightly, which can't be too bad,
and increases atomicity, which can't be bad.


Juno now fails rarely, but not never.
Sometimes it hangs.
Sometimes it has an array access out of bounds.


Assuming m.release() is UnlockMutex.


Renaming the mutex to "giant" really does make
it easier to understand.