fix leak in InitMutex if RegisterFinalCleanup runs out of memory.
Note that this requires locking heap while initLock held,
 which is hopefully ok.
We can't store back the mutex and have it be used until we
are sure we'll be able to clean it up.
(Which also means we can't share the provided Clean function)


It would help if RegisterFinalCleanup provided some
sort of two phase operation where the space could
be reserved first and then the values either
filled in later or the space given up.
That way we could probably unnest the locks.
Consider that shortly.
It seems difficult though esp. to provide "space given up"
unless maybe that is considered rare and the space
just wasted for that case (at least if it is no longer
the last entry in the array).