one version of 64bit file size support, in a branch
features of this version:
  compiler supports assignment INTEGER to LONGINT
  compiler supports various mixed operations
  FOR loop in compiler not changed, but should be
  compiler doesn't quite get assignability correct

ramifications:
  no need for VAL(expr, LONGINT) sprinkled everywhere
    except for a few, due to compiler not quite right
  need for ORD sprinkled
    because don't support LONGINT to INTEGER

There are many bits to decide and this is just combination.

Also, this only has the support for Windows, not Posix.
And the Windows support isn't "real" in that the
compiler makes LONGINT only 32bits, and the few places
that get file sizes from the underlying system,
still truncate to 32bits. However, this diff is "mostly real",
given the "decision bits" it has made.

And one can consider this file list useful, depending
on where we end up.