bring over "AdjustShared" from head
Specifically what this does is when using
older toolsets, it turns building of
anything "shared" into "not shared".


This way these config files can deal
less with the vagaries of older toolsets.


The point being, when mixing these config
files with older tools, only to handle
building cm3.


"old toolset" is defined as "not defining install_symbolic_link"

Perhaps we should have something like cm3 -DBOOT or cm3 -DNO_SHARED
to better encapsulate this.


Note that this is placed in such a way that it doesn't
affect all targets. But probably it should.


Also this disabled building "shared" on Interix, pending
 more investigation. (This is what head does.)