we don't need m3_div, m3_divL, m3_mod, m3_modL We can "just" use gcc's support "directly".
Granted, on many targets, that will generate
calls to __divdi3, __udivdi3, __moddi3, __umoddi3,
but heck, our hand.c code generated such calls as well,
and for platforms for which gcc is not the compiler/linker,
we have libgcc.c.


This *might* be a deoptimization on some platforms, e.g. SOLsun.
That is, if we are going to call a function, better to call their
helpers instead of our libgcc stand-ins.


However it is likely an optimization on
  all 64bit platforms
  all gcc platforms
  all 32bit operations
  leaving the deoptimization only on 32bit non-gcc platforms when using
longint


(Note that m3_divL, m3_modL still used by NT386; m3_div, m3_mod never
were)