another not very useful function in hand.c bites the dust, at least under #ifdef
use bts instruction instead of calling set_singleton
bts's notion of how to index bits happens to line up with set_singleton (assuming test p230 is complete enough)
bts isn't fast, but surely this is better overall due to drastically fewer instructions
  note that set_singleton test cases tend to generate very large code zeroing stuff


I was not able to get coverage of several addressing modes,
so they are left issuing errors and asserts.
That is, among:
  bts mem, reg
  bts reg, imm8
  bts reg, reg
  possibly bts mem, imm8

I was only able to get bts reg, reg to hit.


Still to ponder gcc backend?
I had hoped to get this done in the frontend but became stumped. ?
 It's just a shift and and, should be achievable portably. ?