implementation of direct thread suspend/resume for interix  been sitting on it and almost deleted it
direct suspend/resume can be achieve by writing certain control
codes to the correct files in /procfs

the context of a suspended thread can be gotten by reading a file

However I'd like to at least avoid opening the files repeatedly.

Here we open them in thread create.

I don't remember why Interix can't use the portable code.
It doesn't have SA_SIGINFO but that doesn't prove it.
It is empirically showable that the context is on the stack.
So maybe we just have a single parameter signal handler and a null context?