correct and increase restrictions on extract parameters n and m if constant must be positive (they should probably be CARDINAL or 0..63 instead of INTEGER) if sign_extend, then must be constant and > 1
There is historical code for dealing with non-constant n and sign extension,
and it is apparently clever and correct, except it doesn't work for n = 0.
If we need this to work better, we could insert a runtime check for n = 0,
or use code that handles 0 (see hand.c, at least for now).


Word/Long do not expose extract with sign extension.
The option is available only to m3back clients, ie. m3front,
and it's not hard to imagine that the number of bits it wants
is always constant and never 0.