compare operandPart when comparing stackp
really something larger is bothering me, looking into
the assertion failure when I try to remove unnecessary
instructions from insert/extract

We get into a situation where part of a value is in one
of the correct registers, but hi is in low or vice versa,
and in moving things around, we do the wrong thing.
Like if we want return value in EDX:EAX and the value is
currently in EAX:ECX, we do like:
  XCHG EAX,ECX ; put the low part in EAX where it belongs
  MOV EDX, EAX ; wrong, should be MOV EDX, ECX, or reverse
               ; the two instructions

I don't know that there is a problem here in practise or not,
since I've only seen it so far looking at uncommited changes.