safekeeping: don't save/restore non-volatile registers we don't use This attemps to defer the saving until use, however that appears to be a losing attempt due to branches. Better is to add another pass or some "fixup" so we can do whatever saving is needed in the prologue, but only what is needed.
Compromise will be nop out the unneeded pushes, and omit the pops, of course.
safekeeping: don't save/restore non-volatile registers we don't use
This attemps to defer the saving until use, however that appears to be
a losing attempt due to branches.
Better is to add another pass or some "fixup" so we can do
whatever saving is needed in the prologue, but only what is needed.


Compromise will be nop out the unneeded pushes, and omit the pops, of course.