further decomposition of config files
There are now the following variables:
  m3back_pic, defaults to -fPIC
  m3back_unwind_tables, defaults to -funwind-tables
    (This is in anticipation of using libunwind, some day.)
  m3back_m32, defaults to -m32 or empty, depending on word_size
  m3back_m64, defaults to -m64 or empty, depending on word_size
  m3back_debug, defaults to -gstabus+
  m3back_optimize, defaults to -O2 -Wuninitialized (should be -O3)?

as a result, most platforms don't have to set any flags

if a platform needs a different debug format, it can set just m3back_debug
if a 32bit platform does not work with -m32, it can set m3back_m32 to empty
and so on
  for example I386_SOLARIS sets m3back_debug = "-gstabs"
  for example I386_INTERIX sets m3back_pic = ""
if a platform wants to add other flags, use m3back_flags
  for example -mcpu and -march, though you can also
  piggyback off switches you want to turn off, e.g.:
    MIPS64_OPENBSD:m3back_m64 = "-mabi=64"

Note that this logic is currently in Unix.common and therefore Mingw,
  Cygwin, VMS don't partake (none of this is relevant to native NT386)

add -mpreferred-stack-boundary=2 to I386.common
I386_SOLARIS is seeing a lot of extra code to align stack unnecessarily to 16
e.g. every function call adjusts the stack before the call to align it

still to revisit:
 -munaligned-doubles
 -mno-align-double

I suspect we don't need these with current Target.m3, since
I boosted all the alignments. These might be useful with old compiler
though.