drat: importing a library that imported a system library  only after running configure_system_libs, does not work:
  == package /Users/jay/dev2/cm3/m3-www/proxy ==

  "/cm3.x86/pkg/X11R4/I386_DARWIN/.M3EXPORTS", line 23: quake runtime error:
     Unable to use definition of "SYSTEM_LIBS" from configuration file "/Users/jay/dev2/cm3/m3-sys/cminstall/src/config-no-install/I386_DARWIN": does not define a value for "X11"

for now, therefore, on I386_SOLARIS and PPC64_DARWIN always
 run configure_system_libs up front, slow but ok

on the rest, just do what we were doing -- assuming what was present

This needs more looking into, as for example it was "threatening"
to improve things, e.g. look automatically in known but not
always present locations for stuff, only if asked for.
 (e.g. Motif on Darwin)

If I'm not completely out in left field, then a solution is
make this more complete and have cm3 output
 "configure_system_libs" before "import_sys_lib()"

or something maybe more sophisticated, such as

callers would unconditionally import_sys_lib()
and internally cm3 would call configure_system_libs,
and then if the library isn't present, silently do nothing
for this directory, and anything that imports it
 maybe with a warning
 (actually it would maybe export a stub .M3EXPORTS, again
 with "failed_import_sys_lib")

so that the later importer would get a sensible warning:
 not building foo, because it imports bar, which is dependent on abc,
   which isn't present