GCC 4.2/4.3 require: struct foo GTY()) { int foo; }
GCC 4.5 requires:
struct GTY()) foo { int foo; }

It does not appear to be amenable to #ifdef, since the tool
that scans for the GTY(()) is one of those infamous
not-a-C-parser programs (which is everything except
actual C compilers), so split up the content into separate files.

The annotation syntax on global variables and typedefs appear
to be the same, so those are left alone.

This is yet a no-op, as moving to GCC 4.5 is work in progress.