put div/mod back to long standing and release form:   almost always call a function, unless both parameters known   to be positive or the type is unsigned
  Modula-3 very unfortunately defines div/mod differently than
   pretty much everyone else. Perhaps perhaps the language
   definition can be repaired. It turns out there are several
   reasonable definitions of div and mod. Modula-3 doesn't
   chose the obvious correct one and C doesn't chose the
   obvious incorrect one.

We may yet be able to use the gcc support but it makes me nervous.
It requires at least in a very minimal sense, for the frontend
to support TImode (aka int128_t).

See: http://gcc.gnu.org/ml/gcc/2010-06/msg00647.html
(which is just from me, talking to myself..)