This should cleanup configure -enable-checking a little more, regarding m3_do_fixed_extract, which is often used for integer division, signed and unsigned.
The only change in m3core was a small optimization:
+++ AMD64_DARWIN/RTTipe.ms	2010-06-29 05:59:08.000000000 -0700
@@ -3366,41 +3366,39 @@
 	.stabd	68,0,516
-	movq	-8(%rbp), %rax
-	sarq	%rax
-	movq	%rax, -8(%rbp)
+	sarq	-8(%rbp)

It is not all clear.
The tree ended up with, as configure -enable-checking reports,
two types, the input and output. -enable-checking demands
they be the same. For the left shift, the type doesn't matter.
For the right shift, I believe we want the possibly signed type.