flesh out infrastructure for checking in known/suspected correct assembly and then comparing to current assembly
the intent, if not the code, is:

 in any directory
   m3-sys/m3tests/src/c1/c123
 create
   m3-sys/m3tests/src/c1/c123/expected/I386_LINUX/Main.ms
   m3-sys/m3tests/src/c1/c123/expected/I386_LINUX/Foo.ms
   m3-sys/m3tests/src/c1/c123/expected/AMD64_LINUX/Main.ms
   m3-sys/m3tests/src/c1/c123/expected/AMD64_LINUX/Foo.ms
   etc.

 The name "expected" is a hardcoded part of the interface.
 I'm open to making it something else or additional flexibility,
 but this seems adequate or better to me.

 Any file in expected/TARGET is compared against TARGET.
   Missing/extra files are just ignored.

 Note that this probably works for stdout.pgm and such.

 Future: We should probably allow for:
   m3-sys/m3tests/src/c1/c123/expected/I386/Foo.ms
   m3-sys/m3tests/src/c1/c123/expected/AMD64/Main.ms

 once we confirm that they are often the same.
 e.g. I wouldn't be surprised if they are the same across all
   Solaris, *BSD, Linux, though probably not Darwin.
   Heck, we might even make up something called I386_ELF
   or I386_SYSV.
Cross that bridge later.

If the code does not agree with this description, well,
this description was my intent.

Whether or not creating a directory was worthwhile, it still not clear.

The previous implementation allowed for just:
   m3-sys/m3tests/src/c1/c123/Main.ms-AMD64_LINUX

and then you could either glom everything into Main.m3,
or use multiple test cases.

At the moment I'm thinking of having multiple files but
just one large test case. We'll see.

It's not a big deal either way.