Stripped-down version of Elego ComPact; software developed and used at Elego years ago. I put it under the same copyright as the other Modula-3 sources from DEC and Critical Mass.
Included is package version management based on CVS and build and
version management for sets of packages, called `project management'
in ComPact.

Not included is the general package manager for C/C++/Java/M3, and the
ComPact GUI.

It is contributed in the hope that it may be useful in whole or parts
for the CM3 project or any other users.

The command line interface consists of the programs pkgvm and prjm,
which include complete manuals as well as short usage help:

  pkgvm -help | -man
  prjm  -help | -man

The sources have been successfully compiled with a current cm3 (d5.9.0)
by prjm with

  prjm -D ROOT=$HOME/work/cm3 -buildglobal

The project definition is contained in the file elego/PrjDesc.

To compile everything with the CM3 srcipts, use

  ~/work/cm3/scripts/do-pkg.sh buildship langutils graphicutils tuple depgraph miniscan fsfind fileinfo filepool pkgversions compact-copyright rsrcservice compact-client compactrc pkg_base pkg_vc m3makefile m3dep m3dir m3err m3msh m3ovr pkgvm prjbase prjm

Several things may still be wrong and need to be adapted for use with
the current CM3, but the ComPact tools should be quite flexible and
configurable.

As all free software, this comes without any guarantee that it will
be useful in any way. Use at your own risk.

Olaf Wagner / Elego Software Solutions GmbH, Berlin, Germany, 2010-08-07