Deleting the copied cm3cg file completely breaks the extraction of the m3cc build in the Hudson regression builds.
m3cc should there only be compiled in the cm3-(current-)m3cc jobs,
and not in the -build= and -test- jobs.

This is achieved by setting the PREBUILT_CM3CG environment variable
to a ready-to-use cm3cg.
This in turn is done in the hudson_build_system.sh and hudson_pkg_tests.sh
scripts.

This mechanism should not be used in any other place.

So I added the "and not NOACTION" part here to make it work again,
even if the local m3cc might not be up-to-date.