add detection of AMD64_NETBSD to pylib.py probably fix detection of AMD64_NETBSD in sysinfo.sh (amd64 vs. x86_64) change NetBSDv2_i386 to I386_NETBSD in sysinfo.sh
(Note that cm3 is actually way over-parameterized here.
For example, the only thing the Posix x86 platforms vary
on is the size of jmpbuf. The only think most platforms
vary on are word size, posix or not, jmpbuf size, endian.
This should probably all be moved into config files
and config files only.)